+++ apps/router-e2e/__e2e__/03-server-actions-only/app/_layout.tsx
@@ -1,0 +1,6 @@
+'use client';
+import { Stack } from 'expo-router';
+
+export default function Layout() {
+  return <Stack />;
+}