+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -633,7 +633,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     files: ExportAssetMap;
   }> {
     // NOTE(EvanBacon): This will not support any code elimination since it's a static pass.
-    const {
+    let {
       reactClientReferences: clientBoundaries,
       reactServerReferences: serverActionReferencesInServer,
       cssModules,
@@ -649,7 +649,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     debug('Evaluated client boundaries:', clientBoundaries);
 
     // Run metro bundler and create the JS bundles/source maps.
-    const bundle = await this.legacySinglePageExportBundleAsync(
+    let bundle = await this.legacySinglePageExportBundleAsync(
       {
         ...options,
         clientBoundaries,
@@ -676,13 +676,33 @@ export class MetroBundlerDevServer extends BundlerDevServer {
 
     debug('React server action boundaries from client:', reactServerReferences);
 
-    await this.rscRenderer!.exportServerActionsAsync(
-      {
-        platform: options.platform,
-        entryPoints: [...serverActionReferencesInServer, ...reactServerReferences],
-      },
-      files
+    // When we export the server actions that were imported from the client, we may need to re-bundle the client with the new client boundaries.
+    const { clientBoundaries: nestedClientBoundaries } =
+      await this.rscRenderer!.exportServerActionsAsync(
+        {
+          platform: options.platform,
+          entryPoints: [...serverActionReferencesInServer, ...reactServerReferences],
+        },
+        files
+      );
+
+    const hasUniqueClientBoundaries = nestedClientBoundaries.some(
+      (boundary) => !clientBoundaries.includes(boundary)
     );
+
+    if (hasUniqueClientBoundaries) {
+      debug('Re-bundling client with nested client boundaries:', nestedClientBoundaries);
+      clientBoundaries = [...new Set(clientBoundaries.concat(nestedClientBoundaries))];
+      // Re-bundle the client with the new client boundaries that only exist in server actions that were imported from the client.
+      // Run metro bundler and create the JS bundles/source maps.
+      bundle = await this.legacySinglePageExportBundleAsync(
+        {
+          ...options,
+          clientBoundaries: [...clientBoundaries, ...nestedClientBoundaries],
+        },
+        extraOptions
+      );
+    }
 
     // Inject the global CSS that was imported during the server render.
     bundle.artifacts.push(...cssModules);
@@ -834,7 +854,10 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     const config = getConfig(this.projectRoot, { skipSDKVersionRequirement: true });
     const { exp } = config;
     // NOTE: This will change in the future when it's less experimental, we enable React 19, and turn on more RSC flags by default.
-    const isReactServerComponentsEnabled = !!exp.experiments?.reactServerComponents;
+    const isReactServerComponentsEnabled =
+      !!exp.experiments?.reactServerComponents || !!exp.experiments?.reactServerActions;
+    const isReactServerActionsOnlyEnabled =
+      !exp.experiments?.reactServerComponents && !!exp.experiments?.reactServerActions;
     this.isReactServerComponentsEnabled = isReactServerComponentsEnabled;
 
     const useServerRendering = ['static', 'server'].includes(exp.web?.output ?? '');
@@ -984,6 +1007,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
           rscPath: '/_flight',
           ssrLoadModule: this.ssrLoadModule.bind(this),
           ssrLoadModuleArtifacts: this.metroImportAsArtifactsAsync.bind(this),
+          useClientRouter: isReactServerActionsOnlyEnabled,
         });
         this.rscRenderer = rscMiddleware;
         middleware.use(rscMiddleware.middleware);
@@ -1022,6 +1046,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
           rscPath: '/_flight',
           ssrLoadModule: this.ssrLoadModule.bind(this),
           ssrLoadModuleArtifacts: this.metroImportAsArtifactsAsync.bind(this),
+          useClientRouter: isReactServerActionsOnlyEnabled,
         });
         this.rscRenderer = rscMiddleware;
       }